add `ui/toggle-button` tests
